#  Optimize the ernie inference performance on xpu backend.