<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
Others
### Describe
<!-- Describe what this PR does -->
Inference 使用 xpu backend 推理增加 cache 机制，优化 ernie 模型，ernie medium 模型性能:
  首次推理时间( 65 ms -> 20 ms)，非首次推理时间( 45.07 ms -> 2.5 ms)
具体改动：

-  修复  operator.cc 中关于 enable_cache_runtime_context 和 prepare_data 的逻辑,

-  删除 pre_scope_ = cur_scope 的判断逻辑，该判断逻辑只应用于 enable_cache_runtime_context_ == true 的场景，判断是否重新生成 runtime_context，所以只需判断 runtime_ctx_.get() == nullptr 即可，因为 enable_cache_runtime_context_ 本身就是要复用 runtime_context，上述判断逻辑多余且非必要。
- 区分 transfer_cache_scope 和 new_scope，new_scope 在 enable_cache_runtime_context_ == true 的场景下，runtime_context 中的 variable 变量会使用 new_scope 中的变量代替，因此需要保证，每次 run 的时候，new_scope 必须存在，不能删除，但是对于 cpu 推理的场景，transfer_cache_scope 是不需要的，所以二者需要区分开。
- 增加重新创建 runtime_context 逻辑:
  -  enable_cache_runtime_context 时，程序第一次运行。
  - enable_cache_runtime_context 时，程序非第一次运行，但是运行时 input shape 或者 layout 发生变化。